Fix #29: Part 3 - Instrument most front-end actions.
Let's instrument everything on day one and claw back later.  This sniffs
every action with a truthy `instrument` field, tracking only the event
type.

This immediately exposes a number of places in the App where we're not
fully React-y.  For example, we don't dispatch an action in response to
requesting a new window, and tapping the bookmark star dispatches an
aggregate SET_PAGE_DETAILS action.  I think we'll want to make the
former an action, to facilitate tracking, and the latter a discrete
action.  That separates aggregates; is easier to grep for; enables
fine-grained UI updates; and makes this approach more effective.

Over time, we might grow an explicit user action/command layer rather
than piggy-backing on React messages, which might be too low level.
This is what Atom does.