Make the Node.js process run a Redux loop.
There's much more to write here, but:

* I fix the terminology 'profile command' to refer to an IPC message
  from render process to main process asking the profile service to
  update the canonical profile store.  We differentiate from action
  since commands may be rejected.

* I fix the terminology 'profile diff' to refer to an IPC message from
  main process to render process informing an individual browser window
  that it's view of the profile store needs to change.  We differentiate
  from action to preserve that namespace for the traditional internal
  browser window actions.

This is "full stack Redux" with the Electron IPC layer for transport and
an explicit "command" and "diff" protocol rather than replacing the
entire client Redux state each update.

This handles one small case, the bookmark star.  Functionally, it shows
the star in a disabled state while the page is loading.  (This reveals
some issues with our event handling that I believe are due to iframe and
sub-resource loads.)  It also sets the bookmark star state
optimistically on tap; the canonical state is still propagated from the
profile service over IPC.

It turns out to be much easier to maintain the profile state in the
browser window state, because profile messages almost always update
browser window state.  I left the `state.browserWindow` in place in
anticipation of future (independent) client state reducers.