Transpile all the main process scripts and drop hot-module-loading.
This uses babel to transpile all of the main process scripts and also removes
all of the hot-module support and the web server that it required. All generated
scripts go into the `lib` directory and the app runs from there using file links
which will make using preload scripts much easier.

For the main process scripts we only transpile if there has been a change.

Signed-off-by: Dave Townsend <dtownsend@oxymoronical.com>