Support downloading a custom build of electron. r=nalexander
Drops the electron-prebuilt dependency and instead manually checks the installed
electron version and if incorrect downloads with electron-download. We cache
the download locally to tofino since the global cache could affect other
applications on a developer's machine. Currently this doesn't change where
electron is downloaded from but once we have builds available we can just set
the electron-download options to point to them.

Signed-off-by: Dave Townsend <dtownsend@oxymoronical.com>