Part 1 - Add profileCommands.{send,request}.
profileCommands.request is for making User Agent service requests that
expect responses.  As an example, rather than having the UA cache and
radiate all known URL bar completions, we just respond with the
requested completions.  The UI can then fold the completions into its
cache, which it may manage (drop!) whenever it chooses.  This XHR-like
approach should make it easier to experiment on the front-end, and is
similar to requesting search completions from external services.

The profile command handling in the main process is crufty, and will be
cleaned up eventually.