Part 3 - Track browser sessionId and ancestorId for tabs.
The pattern here is optimistically create a tab in the UI process and
then asynchronously ask the User Agent service to start the
corresponding browsing session.  This means that it's technically
possible to tell the User Agent service to do additional things before
it's allocated a session ID; for now, we null-check the sessionID on the
UI side.  In time, we could drop such requests on the service side if we
prefer.

A small piece of this is renaming the browser window `sessionId` to be
what it really is, which is a scope identifier.  I'm abusing the storage
layer to allocate scope identifiers for now, even though they're a
separate namespace.

The profile command handler remains a dog's breakfast, mostly because we
can both "respond" to a request, and also update the main process state.
Over time, I think we distinguish these two actions, potentially by
moving the "real" User Agent requests (those that touch storage and
should reflect to other clients (start session, bookmark, etc)) to a Web
Socket server and handling the "local" commands (those that are Electron
specific (new window, close window, etc)) separately.  In this model we
might drop the main process state cache entirely; or we might have the
main process be a Web Socket client and listen to updates (that it then
radiates further) itself.