Pre: Don't set <webview hidden>.
Per
https://github.com/electron/electron/blob/master/docs/api/web-view-tag.md#css-styling-notes,
we need to be careful with <webview> visibility.  I observe locally that
webviews don't load when they are hidden.  This suggests we should
differentiate whether the page is the selected page from whether
webviews are active (painting, running JavaScript).  We can hide using
width and height webviews that are active but not selected; and then we
hide using the hidden attribute webviews that are not active.

This patch makes the subsequent patch for loading background tabs work,
but introduces terrible flicker when switching tabs.  We already had
some flicker, so hopefully we can work around this.