Make the main menu open from the hamburger button too. r=bgrins
We cache the bookmarks data so we can rebuild the menu at any time and some
bits are moved around to allow for building menus for minimal OSX, regular
application and the hamburger menu. The hamburger menu has a slightly different
template allowing important options to be at the top level.

Signed-off-by: Dave Townsend <dtownsend@oxymoronical.com>