Show and focus URL bar when creating a new tab. r=linclark. (#521)
This makes more actions per-Page (and require a page ID), and grows the
set of UI state stored per-Page.  This doesn't separate the storage.
Lin Clark has work to do that; I'm not going to bother to do it here,
since it'll be the same for the old userTypedLocation and the new fields
that are now per-Page.

This also shows and focuses the URL bar when creating a new window,
since a new window immediately creates a new tab.  I've elected not to
hide the completion results when tabs are switched; I feel like that's a
separate choice, which we may or may not want.  (Firefox hides them when
you tab away, and doesn't show them when you tab back.)

It's safe to dispatch multiple actions sequentially because we
don't (yet!) have middleware that can delay plain (non-thunk) actions.
We could elect to allow dispatching lists (Arrays) of actions to save
renders and set a pattern for avoiding races in the future, but the
approach would need to account for our thunk middleware.  In spirit,
this is similar to
https://twitter.com/dan_abramov/status/656074974533459968 and
https://github.com/tshelburne/redux-batched-actions.  However, those
approaches don't work with our thunk middleware.