Separate Browser Toolbox from Developer Tools. r=vporof
The former is the Tofino UI (BrowserWindow); the latter is a particular
page displaying web content (<webview> element).

It appears to not be possible to embed (not detach) the Developer Tools
into (or adjacent to) the <webview> element.  It would be nice to do
that, because the current scheme doesn't handle shortcuts well.  In
particular, hotkeys in detached devtools windows aren't recognized,
because these devtools windows aren't identified as BrowserWindows.  We
may be able to register a subset of hotkeys for devtool windows in the
future.