Don't wait for stdout/stderr before exiting
There is a race condition between when the process exits and when the stdout and stderr are drained. As a rule, process.exit() will always exit the process immediately, regardless of whether or not there's still something to be done, even like a console.log; furthermore, logging operations themselves aren't atomic, so performing multiple error logging operations consecutively might result in garbled output if they're not drained in succession. This patch only circumvents the issue, because since we drain the output in build/util.js/webpackBuild anyway, and that's the only place we expect rich long output, we can probably safely refrain from draining stdout and stderr at the top level as well. So what ended up happening was that we were exiting normally, because a forced process.exit(1) wasn't being called since we were waiting for the output to drain.

Signed-off-by: Victor Porof <vporof@mozilla.com>