Add bunyan as a logging framework and use it in all non-content processes. r=victor
Bunyan was mainly chosen in preference to winston because winston doesn't work
in a webpacked environment.

When a moduile first needs the logger it will be configured for the process.
Processes that run in a web context log to the regular console, other processes
log to stdout. The UA and content services also log to a file which helps
debugging issues on windows.

Webpack is used to alias 'logging' to the correct logging module for the
process.