First cut at persisting datoms to SQLite.
This adds the concept of a snapshot (the `datoms` table) and transaction
log (the `transactions` table).  Updating the snapshot is potentially
expensive; writing to the log should not be expensive.  (All for
SQLite.)  Therefore I'm going to separate the two activities and have
the application layer determine when to update the snapshot.

Next steps are to incrementally update the snapshot, to do some stress
testing, and to consider building Datomic API-alikes exposing the
transaction history, etc.