Pre: Use 1 instead of true as placeholder value for page/starred.
The issue is that JavaScript boolean true and false are converted to
integer 1 and 0, respectively, when written to SQLite.  However, we have
no way to convert the integers back to JavaScript booleans when reading
from the database automatically.  We could add type annotations to the
DataScript schema, but I'd like to avoid that for as long as possible.