Pre: Don't process.exit the importing process by default.
I was witnessing incredibly difficult to trace "exit code 2" from Mocha
when *importing* user-agent-service/index.js.  What was happening is
that the import bound process events to process.exit, and
other (unrelated!) tests had unhandled rejections.  This avoids the
unsightly exit, and also updates the exit codes to give a hint of the
issue in the future.  (Exit codes 1 and 2 are valid Mocha exit codes.)