Spawn User Agent Service during tests instead of embedding a UserAgentService.
jsantell and I discussed this on Vidyo, and we agreed that the
"boundary" for the User Agent Service should be a shell
script (bin/user-agent-service) with fixed parameters, provided by an
NPM dependency.  The alternative would be to have the "boundary" be a
JavaScript API allowing to embed a UserAgentService, like we had for the
current tests.  This patch removes that embedding usage from the Tofino
tree, paving the way for externalizing the shell script.